Collapse a list of bidirectional DIAMOND hits

Collapse a list of bidirectional DIAMOND hits

Usage

collapse_bidirectional_hits(blast_inter, compare)

Arguments

blast_inter

A list of data frames containing BLAST/DIAMOND tables with comparisons between target species and outgroups. BLASTp, DIAMOND, or similar programs must be run on processed sequence data as returned by syntenet::process_input().

compare

A 2-column data frame with target species name in column 1, and outgroup species name in column 2. Species names must match names of list elements in blast_inter.

Value

A list of data frames with BLAST/DIAMOND tables as in blast_inter, but with bidirectional hits combined. For instance, if blast_inter contains elements 'spA_spB' and 'spB_spA', these two data frames are combined into a single data frame following the order indicated in outgroups (i.e. 'column1_column2').

Examples

data(blast_list)
blast_inter <- blast_list[c(2,3)]
compare <- data.frame(species = "Olucimarinus", outgroup = "OspRCC809")
chits <- collapse_bidirectional_hits(blast_inter, compare)