Detect interspecies synteny

Detect interspecies synteny

Usage

interspecies_synteny(
  blast_inter = NULL,
  annotation = NULL,
  inter_dir = file.path(tempdir(), "inter"),
  anchors = 5,
  max_gaps = 25,
  is_pairwise = TRUE,
  verbose = FALSE,
  bp_param = BiocParallel::SerialParam(),
  ...
)

Arguments

blast_inter

A list of BLAST/DIAMOND data frames for interspecies comparisons as returned by run_diamond().

annotation

A processed GRangesList or CompressedGRangesList object as returned by process_input().

inter_dir

Path to output directory where .collinearity files will be stored.

anchors

Numeric indicating the minimum required number of genes to call a syntenic block. Default: 5.

max_gaps

Numeric indicating the number of upstream and downstream genes to search for anchors. Default: 25.

is_pairwise

specify if only pairwise blocks should be reported Default: TRUE.

verbose

Logical indicating if log messages should be printed on screen. Default: FALSE.

bp_param

BiocParallel back-end to be used. Default: BiocParallel::SerialParam().

...

Any additional arguments to the MCScanX algorithm. For a complete list of all available options, see the man page of rcpp_mcscanx_file().

Value

Paths to .collinearity files.

Examples

# Load data
data(proteomes)
data(blast_list)
data(annotation)

# Get DIAMOND and processed annotation lists
blast_inter <- blast_list[2]
annotation <- process_input(proteomes, annotation)$annotation

# Detect interspecies synteny
intersyn <- interspecies_synteny(blast_inter, annotation)