Parse .collinearity files obtained with MCScan

The .collinearity files can be obtained with intraspecies_synteny and interspecies_synteny, which execute a native version of the MCScan algorithm.

Usage

parse_collinearity(collinearity_paths = NULL, as = "anchors")

Arguments

collinearity_paths

Character vector of paths to .collinearity files.

as

Character specifying what to extract. One of "anchors" (default), "blocks", or "all".

Value

If as is "anchors", a data frame with variables "Anchor1", and "Anchor2". If as is "blocks", a data frame with variables "Block", "Block_score", "Chr", and "Orientation". If as is "all", a data frame with all aforementioned variables, which indicate:

Block

Numeric, synteny block ID

Block_score

Numeric, score of synteny block.

Chr

Character, query and target chromosome of the synteny block formatted as "&".

Orientation

Character, the orientation of genes within blocks, with "plus" indicating that genes are in the same direction, and "minus" indicating that genes are in opposite directions.

Anchor1

Character, gene ID of anchor 1.

Anchor2

Character, gene ID of anchor 2.

Examples

collinearity_paths <- system.file(
    "extdata", "Scerevisiae.collinearity", package = "syntenet"
)
net <- parse_collinearity(collinearity_paths)