Detect interspecies synteny

Detect interspecies synteny

interspecies_synteny(
  blast_inter = NULL,
  annot_list = NULL,
  inter_dir = file.path(tempdir(), "inter"),
  anchors = 5,
  max_gaps = 25,
  is_pairwise = TRUE,
  verbose = FALSE,
  ...
)

Arguments

blast_inter

A list of BLAST tables for interspecies comparisons.

annot_list

A processed GRangesList or CompressedGRangesList object as returned by process_input().

inter_dir

Output directory.

anchors

Numeric indicating the minimum required number of genes to call a syntenic block. Default: 5.

max_gaps

Numeric indicating the number of upstream and downstream genes to search for anchors. Default: 25.

is_pairwise

specify if only pairwise blocks should be reported Default: TRUE.

verbose

Logical indicating if log messages should be printed on screen. Default: FALSE.

...

Any additional arguments to mcscanx.

Value

Paths to .collinearity files.

Examples

data(blast_list)
data(annotation)
blast_inter <- blast_list[2]
annot_list <- lapply(annotation, function(x) {
    x$gene <- x$gene_id
    return(x)
})
intersyn <- interspecies_synteny(blast_inter, annot_list = annot_list)