rcpp_mcscanx_file

MCSanX provides a clustering module for viewing the relationship of colinear segments in multiple genomes (or heavily redundant genomes). It takes the predicted pairwise segments from dynamic programming (DAGchainer in particular) and then try to build consensus segments from a set of related, overlapping segments.

rcpp_mcscanx_file(
  blast_file,
  gff_file,
  prefix = "out",
  outdir = "",
  match_score = 50L,
  gap_penalty = -1L,
  match_size = 5L,
  e_value = 1e-05,
  max_gaps = 25L,
  overlap_window = 5L,
  is_pairwise = FALSE,
  in_synteny = 0L,
  verbose = FALSE
)

Arguments

blast_file

blast input

gff_file

gff input

prefix

output prefix (default: out)

outdir

output directory (default: "")

match_score

match score (default: 50)

gap_penalty

gap penalty (default: -1)

match_size

match_size (default: 5)

e_value

e_value (default: 1e-5)

max_gaps

max gaps (default: 25)

overlap_window

overlap window (default: 5)

is_pairwise

specify if only pairwise blocks should be reported (default: FALSE)

in_synteny

specify patterns of collinear blocks. 0: intra- and inter-species (default); 1: intra-species; 2: inter-species

verbose

specify if verbose output (default: FALSE)

Value

list

References

Wang et al. (2012) MCScanX: a toolkit for detection and evolutionary analysis of gene synteny and collinearity. Nucleic acids research. 40.7, e49-e49.

Haas et al. (2004) DAGchainer: a tool for mining segmental genome duplications and synteny. Bioinformatics. 20.18 3643-3646.

Author

Kristian K Ullrich